Move RootViewStore to useRootView
